requirements.txt
streamlit
opencv-python
mediapipe
numpy
pillow
matplotlib
pandas